requirements.txt
Flask==3.0.3
pandas==1.3.5
gunicorn
openpyxl
mysql-connector-python
python-dotenv